+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Added support to xc3_viewer for loading multiple `.wimdo` or `.camdo` files.
 
+### Changed
+* Improved accuracy of render state and render pass assignments for Xenoblade X models.
+
 ### Fixed
 * Fixed an issue where some animation files failed to load with `load_animations` for xc3_model.